# Addon Test: Only optimize react deps if applicable in vitest-plugin